ISSUE #1711 - Minor refactor
